Close the review's streaming, anchoring, and reporting gaps
Streamed section headers now get their depth: the terminal event carries
a section_header_levels map clustered over every page's heights, and the
stream proto discloses that in-stream levels are unset. The text-less
float fallback anchors within the region's own column instead of trusting
cross-column top order, with a two-column regression test, and body_order
now has direct coverage. Health and service info report the accelerator
flavor the build actually carries, unknown extensions stop claiming to be
PNG, and the derived-grid guard multiplies in 64 bits so adversarial row
and column counts saturate instead of overflowing into acceptance.

+++ src/docling_map.cpp
@@ -481,8 +481,10 @@ void DoclingMapper::fold_table(const officev1::TableData& table,
       if (cell_bbox(cell.line_rects(), &box)) *out->mutable_bbox() = box;
     }
   }
-  if (table.rows() > 0 && table.columns() > 0
-      && table.rows() * table.columns() <= kMaxGridCells) {
+  // 64-bit product: adversarial row and column counts must saturate the
+  // guard, not overflow it into acceptance.
+  if (table.rows() > 0 && table.columns() > 0 &&
+      static_cast<int64_t>(table.rows()) * table.columns() <= kMaxGridCells) {
     for (int row = 0; row < table.rows(); row++) {
       docv1::TableRow* out_row = data->add_grid();
       for (int column = 0; column < table.columns(); column++) {
